+++ packages/picasso-charts/src/LineChart/LineChart.tsx
@@ -67,6 +67,8 @@ export type LineConfig = Record<
   { color: string; variant?: 'solid' | 'reference' }
 >
 
+export type Domain = [number, number]
+
 export type BaseChartProps = {
   lineConfig: LineConfig
   unit?: string
@@ -77,9 +79,10 @@ export type BaseChartProps = {
   allowTooltipEscapeViewBox?: boolean
   className?: string
   showBottomYAxisLabel?: boolean
-  showEvenYAxisTicks?: boolean
   children?: ReactNode
-  getXAxisTicks?: (orderedChartData: OrderedChartDataPoint[]) => any[]
+  getXAxisTicks?: (orderedChartData: OrderedChartDataPoint[]) => number[]
+  getYAxisTicks?: (domain: Domain) => number[]
+  formatYAxisTick?: (value: number, domain: Domain) => string
 }
 
 export type Props = BaseChartProps & {
@@ -177,6 +180,9 @@ const useStyles = makeStyles<Theme, Props>(styles, {
   name: 'LineChart'
 })
 
+const defaultGetYAxisTicks = (domain: Domain) =>
+  getD3Ticks(domain[0], domain[1], NUMBER_OF_TICKS)
+
 export const LineChart = (props: Props) => {
   const classes = useStyles(props)
   const {
@@ -191,9 +197,10 @@ export const LineChart = (props: Props) => {
     highlights,
     referenceLines,
     showBottomYAxisLabel,
-    showEvenYAxisTicks,
     children,
-    getXAxisTicks = getChartTicks
+    getXAxisTicks = getChartTicks,
+    getYAxisTicks = defaultGetYAxisTicks,
+    formatYAxisTick
   } = props
 
   const yKey = Object.keys(lines)[0]
@@ -228,6 +235,8 @@ export const LineChart = (props: Props) => {
       calculateTooltipPosition(next, tooltipElem, chartElem)
     }
   }
+
+  const yDomain: Domain = [BOTTOM_DOMAIN, topDomain]
 
   return (
     <div
@@ -264,19 +273,21 @@ export const LineChart = (props: Props) => {
           <YAxis
             type='number'
             dataKey={yKey}
-            unit={unit}
-            domain={[BOTTOM_DOMAIN, topDomain]}
+            // re-charts will append unit even if we have a format function, hence it's removed here
+            unit={formatYAxisTick ? undefined : unit}
+            domain={yDomain}
             tickLine={TICK_LINE}
             axisLine={AXIS_LINE}
             interval={0}
-            ticks={
-              showEvenYAxisTicks
-                ? getD3Ticks(BOTTOM_DOMAIN, topDomain, NUMBER_OF_TICKS)
-                : undefined
-            }
+            ticks={getYAxisTicks(yDomain)}
             minTickGap={MIN_TICK_GAP}
             tickMargin={TICK_MARGIN}
             width={Y_AXIS_WIDTH}
+            tickFormatter={
+              formatYAxisTick
+                ? value => formatYAxisTick(value, yDomain)
+                : undefined
+            }
           />
 
           {referenceLineList}
@@ -323,9 +334,7 @@ LineChart.defaultProps = {
   unit: 'd',
   tooltip: false,
   allowTooltipEscapeViewBox: false,
-  xAxisKey: 'x',
-  showBottomYAxisLabel: false,
-  showEvenYAxisTicks: false
+  xAxisKey: 'x'
 }
 
 LineChart.displayName = 'LineChart'